Split off input validation functions from main
